Use DOT Implicit values, removed TODOS from test, PPI change
Change PPI back to 72 from 96. There was a misunderstanding between the
DOT dpi attribute and how to convert DOT's height and width attributes
to appropriate values in Cytoscape.

DOT Implicit values are now used as the base default for the VisualStyle
created for imported networks.


Removed TODOs from TestVisualStyle. They were unnecessary for the tests